[FIO toup] fpga: imx_m4: boot M4
This driver allows a controlled boot of the M4 processor.
When secure mode is enabled (FIOVB), the hash of the firmware in the
FIT image is compared against the value stored in the mmc RPMB
partition (via OPTEE). If they differ, the firmware in QSPI flash will
be upgraded to the new image before booting.

When secure mode is not enabled (!FIOVB), the hash of the firmware in
the FIT image is compared against the value calculated from QSPI flash
(but using the size of the FIT image firmware as the size of the QSPI
image in flash). If they difer, the firmware in QSPI flash will be
upgraded to the new image before booting.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>